requirements.txt
sentence-transformers==2.2.0
chromadb==0.4.24
pdfplumber==0.11.8
Pillow==9.0.0
torch==1.13.0
openai-clip==1.0.1